+++ util/rand.go
@@ -5,7 +5,7 @@ import (
 	"crypto/rand"
 )
 
-func RandByte(len int) []byte {
+func RandByte(len uint64) []byte {
 	b := make([]byte, len)
 	_, _ = rand.Read(b)
 	return b
